[breaking] Use the keyPath as the key when the animator is not additive. (#22)
* [breaking] Use the keyPath as the key when the animator is not additive.

When the animator is not additive we don't actually want our animations to stack on top of one another.

There are some scenarios where stacking absolute animations is desired, e.g. creating a "keyframe" animation composed of a variety of beginTime-adjusted animations, but in practice this is the exception rather than the norm. As such, the API is being adjusted to use the keyPath as the animation key when the animator is not additive.

* More docs.

* Method name.