Add support for customizing the timeScaleFactor as part of a CATransaction.
If an animator is shared across a variety of animations, it can often be preferable to change the animator's behavior only within the context of a single transaction.

For example, if all animations in a transaction should have their duration scaled by a certain amount a client would previously have had to cache the previous timeScaleFactror, set the new one on the animator, and then restore the timeScaleFactor once the animations were added.

The CATransaction API introduced by this change makes it easier for clients to make per-transaction timeScaleFactor changes like so:

```swift
CATransaction.begin()
CATransaction.mdm_setTimeScaleFactor(0.5)

animator.animate(with: timing, to: layer, withValues: [0, 1], keyPath: .rotation)

CATransaction.commit()
```