Always commit the model layer value inside a CATransaction with actions disabled.
If the timing's duration was 0 or the animation was instant we were setting the model layer value without disabling actions, which could potentially result in an animation occurring.

The proper model value setting logic has now been moved to the top of the animate method and the duplicate logic has been removed.