Allow headless CALayers to implicitly animate when using the sharedLayerDelegate.
Return NSNull from actionForLayer:forKey:  tells Core Animation to stop searching for an action provider, essentially disabling the default CALayer implicit animation path.

When using MotionAnimator's ahredLayerDelegate we do not want to change any default behavior, so rather than return NSNull we now return nil. nil indicates to Core Animation that it should continue searching for an action provider. In the case of a headless CALayer, this allows +defaultActionForKey to be invoked.