Add UIKit behavioral tests.
These tests provide a baseline of expectations for how different key paths are expected to animate in UIKit. These tests serve both as educational ("What can I animate with UIKit?") as well as as a measure of our own adherance to the baseline expectations.

For example, as a result of these unit tests it is now clear that we should not be additively animating opacity by default.