Refactor common code from explicit/implicit animation implementations.
Implicit animations are unable to read from the layer because the model layer has already changed by the time the animations get generated. This is problematic when beginFromCurrentState is enabled and the layer needs to extract its initial value.

Prior to this change, the implicit animation logic would attempt to read the same model layer value that had already been written, resulting in a no-op animation.

After this change, the implicit animation logic will cache the prior model/presentation layer values so that they can be used as the animation's initial values.