Don't animate system views with the animator.
Notably, animating the status bar with the animator can cause all of the status bar elements to animate. This is not desirable and deviant from UIKit's implicit animator behavior, but at this time I don't see a way to mimic the same behavior without regressing on our support for animating unhosted layers.